fix: 修复QML不识别qint64类型的问题
- 将Q_PROPERTY类型从qint64改为double
- 将QML属性类型从qint64改为real
- 修复插件无法加载的问题

+++ networkmonitorapplet.cpp
@@ -51,24 +51,24 @@ bool NetworkMonitorApplet::init()
     return DApplet::init();
 }
 
-qint64 NetworkMonitorApplet::downloadSpeed() const
+double NetworkMonitorApplet::downloadSpeed() const
 {
-    return m_downloadSpeed;
+    return static_cast<double>(m_downloadSpeed);
 }
 
-qint64 NetworkMonitorApplet::uploadSpeed() const
+double NetworkMonitorApplet::uploadSpeed() const
 {
-    return m_uploadSpeed;
+    return static_cast<double>(m_uploadSpeed);
 }
 
-qint64 NetworkMonitorApplet::totalDownload() const
+double NetworkMonitorApplet::totalDownload() const
 {
-    return m_totalDownload;
+    return static_cast<double>(m_totalDownload);
 }
 
-qint64 NetworkMonitorApplet::totalUpload() const
+double NetworkMonitorApplet::totalUpload() const
 {
-    return m_totalUpload;
+    return static_cast<double>(m_totalUpload);
 }
 
 QStringList NetworkMonitorApplet::networkInterfaces() const
